Adjust legacy kludges
The actual library intended for use is "js3-mode.el".  Adjust its first
and last lines to mention that name instead of "js3.el".

"js3.el" is a symlink to "js3-mode.el".  Some users might try to load it
using (require 'js3).  Provide that feature in "js3-mode.el".

The files in "lib/" are only intended to be used to generate
"js3-mode.el" and it should therefore not be placed on the
`load-path'. Add a new file "lib/.nosearch", which makes it less likely
that "lib/" accidentally ends up there.