Avoid a warning from the byte-compiler
There are two options available (1) define `js3-mode-error' and its
alias `js3-mode-help' before using them, or (2) quote the function
symbol using `quote' instead of `function' (' instead of #').

We do the latter here.  The result is that the byte-compiler no longer
knows that these function symbols are intended to be used as functions,
so it no longer checks whether these functions are defined, so it no
longer detects that this may not be the case, and does not complain.
That's okay.  The functions are being defined, just after they are being
defined.

Also note that quoting the first symbol in a `defalias' form as a
function is never right.  When `defalias' arguments are being evaluated
the alias (i.e. the function) is not defined yet, that's why we call it.